(cli) Separate cli by features
